+++ src/daf/__init__.py
@@ -11,7 +11,7 @@ from .logging import *
 from .web import *
 from .convert import *
 from .remote import *
-
+from pathlib import Path as __Path
 
 
 from .misc import DOCUMENTATION_MODE
@@ -20,3 +20,4 @@ if DOCUMENTATION_MODE:
 
 
 VERSION = "2.9.0"
+DAF_HOME_PATH = __Path.home().joinpath("./daf/")